Crud directorio y login

+++ src/app/auth/login/login.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { AuthService } from './auth.service';
 import { Router } from '@angular/router';
+import { NgClass, NgIf } from '@angular/common';
+import Swal from 'sweetalert2';
 
 interface LoginData {
   email: string;
@@ -11,7 +13,7 @@ interface LoginData {
 @Component({
   selector: 'app-login',
   standalone: true,
-  imports: [FormsModule],
+  imports: [FormsModule, NgClass, NgIf],
   templateUrl: './login.component.html',
   styleUrl: './login.component.css',
 })
@@ -23,16 +25,38 @@ export class LoginComponent {
     password: '',
   };
 
-  onLogin() {
-    this.authService.login(this.login).subscribe((data) => {
-      if (data.status) {
-        localStorage.setItem('isAuthenticated', 'true');
-        localStorage.setItem('token', data.token);
-        this.router.navigate(['entidad/informacion-general']);
-        // this.router.navigateByUrl('entidad/informacion-general');
-      } else {
-        alert('usuario y/o contraseña incorrecta');
-      }
+  loginError: boolean = false;
+
+  showLoading() {
+    Swal.fire({
+      text: 'Espere un poco por favor.',
+      imageUrl: 'assets/loading.gif',
+      imageWidth: 70,
+      imageHeight: 70,
+      showConfirmButton: false,
+      allowEscapeKey: false,
+      allowOutsideClick: false,
     });
+  }
+
+  closeLoading() {
+    Swal.close();
+  }
+
+  onLogin(loginForm: any) {
+    if (loginForm.valid) {
+      this.showLoading();
+      this.authService.login(this.login).subscribe((data) => {
+        if (data.status) {
+          this.closeLoading();
+          localStorage.setItem('isAuthenticated', 'true');
+          localStorage.setItem('token', data.token);
+          this.router.navigate(['entidad/informacion-general']);
+        } else {
+          this.closeLoading();
+          this.loginError = true;
+        }
+      });
+    }
   }
 }
